+++ example/ios/Flutter/flutter_export_environment.sh
@@ -1,13 +1,13 @@
 #!/bin/sh
 # This is a generated file; do not edit or check into version control.
-export "FLUTTER_ROOT=/usr/local/Caskroom/flutter/1.22.3/flutter"
+export "FLUTTER_ROOT=/Users/karthikponnam/Desktop/SDK/flutter"
 export "FLUTTER_APPLICATION_PATH=/Users/karthikponnam/Desktop/my/plugins/FlutterToast/example"
-export "FLUTTER_TARGET=/Users/karthikponnam/Desktop/my/plugins/FlutterToast/example/lib/main.dart"
+export "FLUTTER_TARGET=lib/main.dart"
 export "FLUTTER_BUILD_DIR=build"
 export "SYMROOT=${SOURCE_ROOT}/../build/ios"
 export "FLUTTER_BUILD_NAME=1.0.0"
 export "FLUTTER_BUILD_NUMBER=1"
 export "DART_OBFUSCATION=false"
-export "TRACK_WIDGET_CREATION=true"
+export "TRACK_WIDGET_CREATION=false"
 export "TREE_SHAKE_ICONS=false"
-export "PACKAGE_CONFIG=/Users/karthikponnam/Desktop/my/plugins/FlutterToast/example/.dart_tool/package_config.json"
+export "PACKAGE_CONFIG=.packages"