+++ example/lib/toast_context.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:fluttertoast_example/main.dart';
 import 'package:flutter/material.dart';
 import 'package:fluttertoast/fluttertoast.dart';